rename TS modules to use kebab-case for consistency
