+++ src/lib/chart/index.ts
@@ -1,5 +1,5 @@
-export * from './quoteFeed';
-export * from './timeBucketConverters';
+export * from './quote-feed';
+export * from './time-bucket-converters';
 export { default as ChartActivityTracker } from './ChartActivityTracker';
 export { default as ChartIQ } from './ChartIQ.svelte';
 export { default as ChartLinker } from './ChartLinker';
